+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Removed Explicit minor versioning of dependencies [#3](https://github.com/as1100k/bevy-discord/pull/3)
 - Disabled default features for `reqwest` [#3](https://github.com/as1100k/bevy-discord/pull/3)
+- Improved Documentation
 
 ### Deprecated
 - module `webhook`