+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.4.0] - Unreleased
 
 ## Removed
-- `webhook` module
+- `webhook` module and feature
 
 ## Changed
 - Add `DiscordHttpResource` once `BReadyEvent` is emitted in `DiscordBotPlugin`